pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "forge"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "jsonschema>=4.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "lark-oapi>=1.3",
    "anthropic>=0.25",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0",
    "pymysql>=1.1",
    "psycopg2-binary>=2.9",
    "jinja2>=3.1",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
forge = "forge.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["forge*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
